Struct evm_adapters::sputnik::cheatcodes::memory_stackstate_owned::MemoryStackStateOwned

pub struct MemoryStackStateOwned<'config, B> {

    pub backend: B,
    pub substate: MemoryStackSubstate<'config>,
    pub trace_enabled: bool,
    pub call_index: usize,
    pub trace_index: usize,
    pub traces: Vec<CallTraceArena>,
    pub expected_revert: Option<Vec<u8>>,
    pub next_msg_sender: Option<H160>,
    pub msg_sender: Option<(H160, H160, usize)>,
    pub accesses: Option<RecordAccess>,
    pub all_logs: Vec<String>,
    pub expected_emits: Vec<ExpectedEmit>,
    pub debug_enabled: bool,
    pub debug_steps: Vec<DebugArena>,

}

This struct implementation is copied from upstream and modified to own the Backend type.

We had to copy it so that we can modify the Stack’s internal backend, because the upstream MemoryStackState only has an immutable reference to Backend which does not allow us to do so.

Fields
backend: Bsubstate: MemoryStackSubstate<'config>trace_enabled: boolcall_index: usizetrace_index: usizetraces: Vec<CallTraceArena>expected_revert: Option<Vec<u8>>next_msg_sender: Option<H160>msg_sender: Option<(H160, H160, usize)>accesses: Option<RecordAccess>all_logs: Vec<String>expected_emits: Vec<ExpectedEmit>debug_enabled: booldebug_steps: Vec<DebugArena>
Implementations

impl<'config, B: Backend> MemoryStackStateOwned<'config, B>

pub fn deposit(&mut self, address: H160, value: U256)

pub fn increment_call_index(&mut self)

pub fn trace_mut(&mut self) -> &mut CallTraceArena

pub fn debug_mut(&mut self) -> &mut DebugArena

pub fn trace(&self) -> &CallTraceArena

pub fn reset_traces(&mut self)

impl<'config, B: Backend> MemoryStackStateOwned<'config, B>

pub fn new(
metadata: StackSubstateMetadata<'config>,
backend: B,
trace_enabled: bool,
debug_enabled: bool
) -> Self