Module evm_adapters::sputnik::helpers

Functions

fuzzvm

Instantiates a FuzzedExecutor over provided Sputnik EVM

new_backend

new_vicinity

vm

Instantiates a Sputnik EVM with enabled cheatcodes + FFI and a simple non-forking in memory backend and tracing disabled

vm_debug

Instantiates a Sputnik EVM with enabled cheatcodes + FFI and a simple non-forking in memory backend and debug enabled, and tracing disabled

vm_no_limit

Instantiates a Sputnik EVM with enabled cheatcodes + FFI and a simple non-forking in memory backend and tracing disabled, and no contract size limit

vm_tracing

Instantiates a Sputnik EVM with enabled cheatcodes + FFI and a simple non-forking in memory backend and tracing enabled

Type Definitions

TestSputnikVM